requirements.txt
python-bale-bot
requests
